# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a lofi music streaming server with community features.",
        "title": "Lofi Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/auth/invitations": {
            "get": {
                "description": "List all pending playlist sharing invitations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get pending invitations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PlaylistShare"
                            }
                        }
                    }
                }
            }
        },
        "/auth/invitations/{id}": {
            "put": {
                "description": "Accept or reject a playlist sharing invitation",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Update invitation status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Share ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status: accepted or rejected",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateShareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stations": {
            "get": {
                "description": "List all lofi music stations (global and user's private/shared ones)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Get all stations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Station"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new lofi music station with a playlist. If authenticated, it will be private.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Create a new station",
                "parameters": [
                    {
                        "description": "Station details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateStationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Station"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stations/{id}": {
            "put": {
                "description": "Update the details and playlist of an existing station. Requires ownership.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Update a station",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Station ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated station details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateStationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Station"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Station not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a station from the list. Requires ownership.",
                "tags": [
                    "stations"
                ],
                "summary": "Delete a station",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Station ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Station not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stations/{id}/share": {
            "post": {
                "description": "Invite another user to access your private playlist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Share playlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Station ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User email to share with",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SharePlaylistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation sent",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateStationRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "playlist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Track"
                    }
                }
            }
        },
        "models.PlaylistShare": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "stationId": {
                    "type": "string"
                },
                "stationName": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, accepted, rejected",
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.SharePlaylistRequest": {
            "type": "object",
            "properties": {
                "userEmail": {
                    "type": "string"
                }
            }
        },
        "models.Station": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "description": "nil means global/system station",
                    "type": "integer"
                },
                "playlist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Track"
                    }
                }
            }
        },
        "models.Track": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "description": "YouTube Video ID",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.UpdateShareRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "accepted, rejected",
                    "type": "string"
                }
            }
        }
    }
}